Remove old duplicate equipment folder, fix imports
When we merged equipment-wip-2 into stage, we forgot that previously
we had named our equipment folder in frontend 'equipment-reservation'.
There were no conflicts due to this, but it caused import errors for
the frontend. We deleted the duplicate folder, updated our imports in
app-routing module, and everything was back to working.